Add more data to struct proc and internal design doc
I added a pointer to the parent process, a counter for the number of children
the process has made, and an array of struct proc pointers to children of
length MAX_CHILDS, currently set to 32.

There are a couple of design decisions being made: having a constant maximum
number of children per process and using an array of struct proc pointers to
hold the child processes. Additionally, there are no explicit parent or child
PIDs, these can be accessed by dereferencing parent or child struct proc
pointers.