Implement `sys_fork()` and needed infrastructure
This is my first swing at implementing the `sys_fork()` call. The call
itself creates a new `struct proc` and then copies in the address space
and various other attributes of its parent `proc`. Then the syscall clones
the trapframe of the parent (caller) and uses `thread_fork()` to run a
new thread, using the helper function `enter_forked_process()` as its target.
This process just changes the `v0` register to `0` to have the correct
return value, adjusts `a0` to `0` to signal success, and increments `epc`
by four to move the program along. Then it switches the trapframe back to
usermode, which *should* get the whole thing going again.