Flesh out `sys_fork()` and related infrastructure
I modified `proc_create()` so that it adds the new proc to the global proc
table by scanning through it (still not O(1) time, Aidan), then assigns it
a PID equal to the index of the process in the global table. We need to
remember to NULL out that spot in `proc_destroy()`. I also changed
`proc_create()` to not be `static` anymore and added it to `proc.h`. This
allows us to use it in `sys_fork()`, which we now do.

I changed `proc_bootstrap()` to init the global proc table at the beginning and
also to init a new global proc table lock (`GPT_lock` defined in `proc.h`).

Most significant are my changes to `sys_fork()`. I changed it to use
`proc_create()` and to add the new process as a child to the old process.